Track update failed events
