+++ WooCommerce/Classes/ViewModels/Booking Details/BookingDetailsViewModel+Analytics.swift
@@ -1,3 +1,4 @@
+import protocol WooFoundationCore.WooAnalyticsEventPropertyType
 import enum Networking.BookingAttendanceStatus
 
 extension WooAnalyticsEvent {
@@ -28,5 +29,29 @@ extension WooAnalyticsEvent {
         static func bookingViewLinkedOrderTapped() -> WooAnalyticsEvent {
             WooAnalyticsEvent(statName: .bookingViewLinkedOrderTapped)
         }
+
+        static func failedToUpdateBookingDetails(action: Action, error: Error) -> WooAnalyticsEvent {
+            var properties: [String: WooAnalyticsEventPropertyType] = [
+                Properties.action: action.rawValue
+            ]
+            properties += error.analyticsProperties
+            return  WooAnalyticsEvent(statName: .bookingListFailedToFetchBookings,
+                                      properties: properties)
+        }
+    }
+}
+
+fileprivate extension WooAnalyticsEvent.BookingsDetail {
+    enum Properties {
+        static let bookingStatus = "booking_status"
+        static let action = "action"
+    }
+}
+
+extension WooAnalyticsEvent.BookingsDetail {
+    enum Action: String {
+        case cancelBooking = "cancel_booking"
+        case updateAttendance = "update_attendance"
+        case markAsPaid = "mark_as_paid"
     }
 }
